Sparky2: check in hardware
These are the hardware files for Sparky2. An open source
flight controller that uses an STM32F4 processor, an
MPU9250 mag, gyro, accel chip, and an MS5611 pressure
sensor. It also has three analog inputs, CAN bus support,
multiple serial and I2C ports, and a flexible receiver
port that is capable of PPM, SBUS, Spektrum, and HOTT.

As the license in the hardware directory states, this
hardware is CC-BY-SA-NC. That means it is illegal to
manufacture this hardware for sale without authorization
from the designer.