targets: remove colibri target and board plugin
Things that would be nice to do, but are not in scope for this PR, and
optional for release (so not even worth a subsidiary issue):

* WOuld be nice to make quanton board plugin know about existing colibri
board id and USB id-- we have no mechanism for multiple IDs in GCS, so
it requires some duplication.
* If so, need to teach upgrade to replace the loader, too, with one that
can support a full quanton image.

Only 2 users of Colibri ever on dR; and I'm not sure they stayed around.